+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+### Changed (v4.0.3)
+
+## 2019-04-12
+
+- Correctly return image name on photos taken with camera on Android
+
 ### Changed (v4.0.2)
 
 ## 2019-04-03